backend/requirements.txt
# T003: Backend dependencies for Todo AI Chatbot (Phase 3)
# Spec Reference: plan.md Dependencies section

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database ORM
sqlmodel>=0.0.14
psycopg2-binary>=2.9.9

# JWT Authentication
PyJWT[crypto]>=2.8.0

# Environment Configuration
python-dotenv>=1.0.0

# MCP and AI Frameworks
mcp>=0.1.0
openai>=1.3.5
openai-agents>=0.0.7

# Dapr Integration (Phase V)
dapr>=1.14.0
dapr-ext-fastapi>=1.14.0

# Date Utilities (Phase V - recurring tasks)
python-dateutil>=2.9.0

# Testing
pytest>=8.0.0
httpx>=0.26.0
pytest-asyncio>=0.23.0
